Add oso

+++ src/todos/delete.js
@@ -2,15 +2,16 @@
 
 const { User } = require('../User');
 const { deleteTodo, getTodo } = require('../db');
-const { error, success } = require('../helpers');
+const { error, may, success } = require('../helpers');
 
 module.exports.delete = async (event, _context, cb) => {
   try {
-    const _user = User.fromEvent(event);
+    const user = User.fromEvent(event);
     const { id } = event.pathParameters;
     const todo = await getTodo(id);
 
-    // TODO: authorize access.
+    const authorized = await may(user, 'delete', todo);
+    if (!authorized) return error(cb, { statusCode: 403 });
 
     await deleteTodo(id);
     success(cb, todo);
